refactor: Merge types/docs from testing/auth/api_key_validator.pyi
- Merged type hints and docstrings from crudclient/testing/auth/api_key_validator.pyi into crudclient/testing/auth/api_key_validator.py.

- Deleted crudclient/testing/auth/api_key_validator.pyi.
